requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
python-decouple
dj_config_url
django-jazzmin
django-cors-headers
pyngrok==5.0.5
